Moves data providers for children domicile transfer
