+++ libs/application/data-providers/src/index.ts
@@ -1,5 +1,3 @@
 export { ExampleFails } from './providers/ExampleFails'
 export { ExampleSucceeds } from './providers/ExampleSucceeds'
 export { UserProfileProvider } from './providers/UserProfileProvider'
-export { ChildrenNationalRegistryProvider } from './providers/ChildrenNationalRegistryProvider'
-export { ParentNationalRegistryProvider } from './providers/ParentNationalRegistryProvider'
