+++ libs/application/templates/children-domicile-transfer/src/dataProviders/index.ts
@@ -1,4 +1,2 @@
-export {
-  ChildrenNationalRegistryProvider,
-  ParentNationalRegistryProvider
-} from '@island.is/application/data-providers'
+export { ChildrenNationalRegistryProvider } from './ChildrenNationalRegistryProvider'
+export { ParentNationalRegistryProvider } from './ParentNationalRegistryProvider'
